+++ server/src/main/java/nl/tudelft/oopp/demo/services/MenuService.java
@@ -1,5 +1,14 @@
 package nl.tudelft.oopp.demo.services;
 
+import static nl.tudelft.oopp.demo.config.Constants.ADDED;
+import static nl.tudelft.oopp.demo.config.Constants.EXECUTED;
+import static nl.tudelft.oopp.demo.config.Constants.MENU_ALREADY_EXISTS;
+import static nl.tudelft.oopp.demo.config.Constants.MENU_NOT_FOUND;
+import static nl.tudelft.oopp.demo.config.Constants.RESTAURANT_NOT_FOUND;
+import static nl.tudelft.oopp.demo.config.Constants.WRONG_CREDENTIALS;
+
+import java.net.URLDecoder;
+import java.nio.charset.StandardCharsets;
 import java.util.List;
 import java.util.Optional;
 
@@ -7,7 +16,9 @@ import nl.tudelft.oopp.demo.entities.Menu;
 import nl.tudelft.oopp.demo.entities.Restaurant;
 import nl.tudelft.oopp.demo.repositories.MenuRepository;
 import nl.tudelft.oopp.demo.repositories.RestaurantRepository;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 /**
@@ -31,15 +42,22 @@ public class MenuService {
     public int add(String name, int restaurantId) {
         Optional<Restaurant> optionalRestaurant = restaurantRepository.findById(restaurantId);
         if (optionalRestaurant.isEmpty()) {
-            return 428;
+            return RESTAURANT_NOT_FOUND;
         }
         Restaurant restaurant = optionalRestaurant.get();
+        if (menuRepository.findByRestaurantId(restaurantId) != null) {
+            return MENU_ALREADY_EXISTS;
+        }
+
+        if (RestaurantService.noPermissions(SecurityContextHolder.getContext(), restaurant)) {
+            return WRONG_CREDENTIALS;
+        }
 
         Menu menu = new Menu();
         menu.setName(name);
         menu.setRestaurant(restaurant);
         menuRepository.save(menu);
-        return 201;
+        return ADDED;
     }
 
     /**
@@ -49,10 +67,13 @@ public class MenuService {
      */
     public int delete(int id) {
         if (!menuRepository.existsById(id)) {
-            return 429;
+            return MENU_NOT_FOUND;
+        }
+        if (RestaurantService.noPermissions(SecurityContextHolder.getContext(), menuRepository.getOne(id).getRestaurant())) {
+            return WRONG_CREDENTIALS;
         }
         menuRepository.deleteById(id);
-        return 200;
+        return EXECUTED;
     }
 
     /**
@@ -70,5 +91,39 @@ public class MenuService {
      */
     public Menu find(int id) {
         return menuRepository.findById(id).orElse(null);
+    }
+
+    /**
+     * Finds a menu with a certain name
+     * @param name menu name
+     * @return menu
+     */
+    public Menu find(String name) {
+        return menuRepository.findByName(URLDecoder.decode(name, StandardCharsets.UTF_8));
+    }
+
+    /**
+     * Finds a menu with a certain restaurant id
+     * @param restaurantId restaurant id
+     * @return menu
+     */
+    public Menu findRestaurant(int restaurantId) {
+        return menuRepository.findByRestaurantId(restaurantId);
+    }
+
+    /**
+     * Updates the name of a menu
+     * @param menuId menu id
+     * @param name the new name for a menu
+     * @return error code
+     */
+    public int changeMenuName(int menuId, String name) {
+        if (menuRepository.findById(menuId).isEmpty()) {
+            return MENU_NOT_FOUND;
+        }
+        Menu menu = menuRepository.findById(menuId).get();
+        menu.setName(name);
+        menuRepository.save(menu);
+        return EXECUTED;
     }
 }
